azzyt-okx-main/requirements.txt
okx>=2.1.0
pandas>=1.5.0
scikit-learn>=1.5.2
numpy>=1.23.0
ccxt
ta==0.11.0
tensorflow==2.17.0
python-dotenv==1.0.1